move etherscan api in explorer folder
