+++ libs/coin-evm/src/api/explorer/index.ts
@@ -1,0 +1,18 @@
+import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
+import etherscanLikeApi from "./etherscan";
+
+/**
+ * Switch to select one of the compatible explorer
+ */
+export const getExplorerApi = (currency: CryptoCurrency) => {
+  const apiType = currency.ethereumLikeInfo?.explorer?.type;
+
+  switch (apiType) {
+    case "etherscan":
+    case "blockscout":
+      return etherscanLikeApi;
+
+    default:
+      throw new Error("API type not supported");
+  }
+};
