+++ libs/coin-evm/src/synchronization.ts
@@ -14,25 +14,10 @@ import {
 import { log } from "@ledgerhq/logs";
 import { CryptoCurrency, TokenCurrency } from "@ledgerhq/types-cryptoassets";
 import { Account, Operation, SubAccount } from "@ledgerhq/types-live";
-import etherscanLikeApi from "./api/etherscan";
+
 import { getBalanceAndBlock, getBlock, getTokenBalance, getTransaction } from "./api/rpc";
 import { getSyncHash, mergeSubAccounts } from "./logic";
-
-/**
- * Switch to select one of the compatible explorer
- */
-const getExplorerApi = (currency: CryptoCurrency) => {
-  const apiType = currency.ethereumLikeInfo?.explorer?.type;
-
-  switch (apiType) {
-    case "etherscan":
-    case "blockscout":
-      return etherscanLikeApi;
-
-    default:
-      throw new Error("API type not supported");
-  }
-};
+import { getExplorerApi } from "./api/explorer";
 
 /**
  * Main synchronization process
@@ -214,12 +199,10 @@ export const getOperationStatus = async (
   op: Operation,
 ): Promise<Operation | null> => {
   try {
-    const {
-      blockNumber: blockHeight,
-      blockHash,
-      timestamp,
-      nonce,
-    } = await getTransaction(currency, op.hash);
+    const { blockNumber: blockHeight, blockHash, timestamp, nonce } = await getTransaction(
+      currency,
+      op.hash,
+    );
 
     if (!blockHeight) {
       throw new Error("getOperationStatus: Transaction has no block");
